refactor(render): one base for the seven family programs
Each family program held a GeometryProgram and delegated the same ten
methods to it, word for word, with compiling ahead written six times
over: adding a method to the geometry program took seven edits, and a
family that forgot one silently took the interface default, compiling
and drawing while paying shaderc on the render thread at its first draw.
The delegations live once on FamilyProgram now; what a family keeps is
what differs, how its geometry is placed and what it is drawn with. The
terrain, the one family that must not compile ahead, says so where it
stands instead of by omission, and the far terrain keeps its guard
over the base road. Behaviour unchanged.
