+++ common/src/main/java/dev/vitrail/render/SkyProgram.java
@@ -4,17 +4,12 @@ import dev.vitrail.glsl.PackProgram;
 import dev.vitrail.glsl.SkyVertex;
 import dev.vitrail.pack.target.ChainPlan;
 import dev.vitrail.pack.target.TargetPlan;
-import dev.vitrail.uniform.WorldState;
 import dev.vitrail.Vitrail;
 
 import com.mojang.blaze3d.pipeline.RenderPipeline;
 import com.mojang.blaze3d.systems.GpuDevice;
-import com.mojang.blaze3d.systems.RenderPass;
-import com.mojang.blaze3d.systems.RenderPassDescriptor;
 import com.mojang.blaze3d.textures.GpuSampler;
 import com.mojang.blaze3d.textures.GpuTextureView;
-import com.mojang.blaze3d.vulkan.VulkanDevice;
-import com.mojang.blaze3d.vulkan.glsl.GlslCompiler;
 
 import org.joml.Matrix4fc;
 import org.joml.Vector4fc;
@@ -44,7 +39,7 @@ import java.util.Set;
  * Sodium's mixin push twenty bytes of region offset into the layout, and the sky is the game's
  * geometry: it has no region and no push constants, and borrowing the word would push them anyway.
  */
-final class SkyProgram implements DumpedProgram {
+final class SkyProgram extends FamilyProgram {
 
 	/** What the log calls this geometry, one word in the middle of a sentence. */
 	private static final String FAMILY = "sky";
@@ -52,10 +47,8 @@ final class SkyProgram implements DumpedProgram {
 	/** Ours, and deliberately without the word that turns push constants on. See the class comment. */
 	private static final String NAMESPACE = Vitrail.MOD_ID;
 
-	private final GeometryProgram body;
-
 	private SkyProgram(GeometryProgram body) {
-		this.body = body;
+		super(body);
 	}
 
 	/**
@@ -137,52 +130,6 @@ final class SkyProgram implements DumpedProgram {
 	}
 
 	/**
-	 * The pass this program is drawn into, or null to leave the renderer the one it meant to open.
-	 *
-	 * @see GeometryProgram#descriptor
-	 */
-	RenderPassDescriptor descriptor(GpuTextureView colour, GpuTextureView depth) {
-		return this.body.descriptor(colour, depth);
-	}
-
-	/**
-	 * Binds this program's block and every sampler it declares, inside the pass just opened.
-	 *
-	 * @see GeometryProgram#bind
-	 */
-	void bind(RenderPass pass) {
-		this.body.bind(pass);
-	}
-
-	/** @see GeometryProgram#compile */
-	@Override
-	public boolean compile(GpuDevice device) {
-		return this.body.compile(device);
-	}
-
-	/** @see GeometryProgram#compiled */
-	@Override
-	public boolean compiled() {
-		return this.body.compiled();
-	}
-
-	/** @see GeometryProgram#forgetCompiled */
-	@Override
-	public void forgetCompiled() {
-		this.body.forgetCompiled();
-	}
-
-	@Override
-	public boolean warmAhead(VulkanDevice device, GlslCompiler compiler) {
-		return this.body.warmAhead(device, compiler);
-	}
-
-	@Override
-	public void discardAhead() {
-		this.body.discardAhead();
-	}
-
-	/**
 	 * Whether the pipeline a pass has bound is this program's.
 	 *
 	 * @see GeometryProgram#owns
@@ -191,39 +138,4 @@ final class SkyProgram implements DumpedProgram {
 		return this.body.owns(bound);
 	}
 
-	/** @see GeometryProgram#decoded */
-	@Override
-	public String decoded(WorldState world) {
-		return this.body.decoded(world);
-	}
-
-	/** @see GeometryProgram#path */
-	@Override
-	public String path() {
-		return this.body.path();
-	}
-
-	/** @see GeometryProgram#label */
-	@Override
-	public String label() {
-		return this.body.label();
-	}
-
-	/**
-	 * Rotates the ring buffer, once the frame's draw has been recorded.
-	 *
-	 * @see GeometryProgram#rotate
-	 */
-	void rotate() {
-		this.body.rotate();
-	}
-
-	/**
-	 * Closes this program's block and the placeholder textures it made.
-	 *
-	 * @see GeometryProgram#release
-	 */
-	void release() {
-		this.body.release();
-	}
 }
